Change the tweet display on Slack to include highlighting of trigger words.
Words that have lead to a score are written in *bold*.
Note that there is a minor change in the scoring: since we match
per word of the tweet now, we can no longer operate with "røde kors"
as search term. The two words are counted and treated individually.