Add an extra threshold that determines when to trigger a channel alert.
There have been quite a few false positives where a single trigger word
caused a tweet to be posted to slack, including a channel alert. Therefore
the alert was disabled, which is kind of against the purpose.
Now we have separate thresholds to steer posting (AZTWITTERSAR_MINSCORE)
and alerting (AZTWITTERSAR_MINSCORE_ALERT). The latter must be higher than
the former, otherwise it will not have any effect.